--- src/entities/Patient/components/InfoBoxEl/index.ts
@@ -1,1 +1,0 @@
-export * from "./InfoBoxEl";
