--- src/entities/Patient/components/InfoNameBlockEl/index.ts
@@ -1,1 +1,0 @@
-export * from "./InfoNameBlockEl";
