--- src/entities/Patient/components/SearchListEl/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchListEl";
