--- src/entities/Patient/components/SearchModalFormEl/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchModalFormEl";
