--- src/entities/Patient/components/SearchModalHeaderEl/SearchModalHeaderEl.styled.tsx
@@ -1,22 +1,0 @@
-import styled from "@emotion/styled";
-
-export const ModalHeader = styled.div`
-  display: flex;
-  align-items: center;
-
-  width: 100%;
-
-  padding-top: ${props => props.theme.spacing(4)};
-  padding-bottom: ${props => props.theme.spacing(4)};
-  padding-right: ${props => props.theme.spacing(6)};
-  padding-left: ${props => props.theme.spacing(6)};
-  border-bottom: ${props => props.theme.border};
-`;
-
-export const ModalTitle = styled.h3`
-  font-weight: 600;
-  font-size: 32px;
-  line-height: 1.13;
-
-  margin-right: auto;
-`;
