--- src/entities/Patient/components/SearchModalHeaderEl/index.ts
@@ -1,1 +1,0 @@
-export * from "./SearchModalHeaderEl";
