+++ src/entities/Patient/components/index.ts
@@ -1,8 +1,4 @@
 export * from "./Patient";
 export * from "./InfoEl";
-export * from "./InfoNameBlockEl";
-export * from "./InfoBoxEl";
 export * from "./SearchEl";
-export * from "./SearchModalHeaderEl";
-export * from "./SearchModalFormEl";
-export * from "./SearchListEl";
+export * from "./SearchModalEl";
