+++ src/entities/Patient/helpers/index.ts
@@ -1,2 +1,1 @@
-export * from "./showBirthDate";
 export * from "./patientInfo";
