+++ src/entities/Patient/redux/patientsSlice.ts
@@ -2,7 +2,6 @@ import { createSlice, PayloadAction } from "@reduxjs/toolkit";
 import { IPatient } from "entities/Patient";
 
 interface IPatientsState {
-  patients: IPatient[];
   patient: IPatient;
   isActiveInfo: boolean;
   modalPatient: boolean;
@@ -26,7 +25,6 @@ const initialState: IPatientsState = {
       postcode: "",
     },
   },
-  patients: [],
   isActiveInfo: false,
   modalPatient: false,
 };
@@ -41,12 +39,6 @@ const patientsListSlice = createSlice({
     deletePatient(state) {
       state.patient = initialState.patient;
     },
-    addPatientsList(state, action: PayloadAction<IPatient[]>) {
-      state.patients = action.payload;
-    },
-    resetPatientsList(state) {
-      state.patients = initialState.patients;
-    },
     fetchIsActiveInfo(state, action: PayloadAction<boolean>) {
       state.isActiveInfo = action.payload;
     },
@@ -59,8 +51,6 @@ const patientsListSlice = createSlice({
 export const {
   addPatient,
   deletePatient,
-  addPatientsList,
-  resetPatientsList,
   fetchIsActiveInfo,
   fetchPatientsModal,
 } = patientsListSlice.actions;
